+++ packages/terminal/src/components/ui/switch.tsx
@@ -4,7 +4,7 @@
  */
 
 import * as React from "react";
-import { Switch as SwitchPrimitive } from "radix-ui";
+import * as SwitchPrimitive from "@radix-ui/react-switch";
 import { cn } from "@/lib/utils";
 
 function Switch({